requirements.txt
python-telegram-bot==13.14
google-cloud-secret-manager==2.8.0